+++ pkg/sectionindex/sectionInfo.go
@@ -2,16 +2,16 @@ package sectionindex
 
 type SectionInfo struct {
 	SeriesSmooth int32
-	StartSeq  int64
+	StartSeq  int32
 	Sign      int8
-	Height    float64
-	Width     int64
-	PrevSeq   int64
+	Height    float32
+	Width     int32
+	PrevSeq   int32
 }
 
 type SectionInfoKey struct {
 	SeriesSmooth int32
-	StartSeq int64
+	StartSeq int32
 }
 
 func (si *SectionInfo) getKey() SectionInfoKey {
